messageHandler/requirements.txt
SpeechRecognition
pydub
langchain-openai
google-generativeai
groq
mysql-connector-python
python-dotenv
sqlalchemy
pyTelegramBotAPI
flask
flask_cors